requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0

# Flask web framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-WTF==1.2.1
gunicorn==21.2.0

# Google AI/Gemini
google-generativeai>=0.8.0
google-adk>=0.1.0

# Additional dependencies for agents
firecrawl-py==1.0.0
faiss-cpu>=1.8.0
sentence-transformers>=2.7.0
beautifulsoup4==4.12.2

# Data processing
numpy>=1.26.0
# pandas>=2.2.0

# Utilities
rich==13.7.0
typer==0.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-flask==1.3.0
